LN ConnectPage refactoring
