+++ app/components/views/LNPage/ConnectPage/CreateLNWallet/CreateLNWallet.module.css
@@ -1,0 +1,38 @@
+.checkbox {
+  max-width: 420px;
+}
+
+.connectOpt {
+  flex-wrap: wrap;
+  display: flex;
+  padding-top: 18px;
+}
+
+.label {
+  display: inline-block;
+  margin-top: 8px;
+  margin-right: 14px;
+  width: 125px;
+}
+
+.accountSelection {
+  display: grid;
+  grid-row-gap: 10px;
+  justify-items: right;
+}
+
+.accountList {
+  width: 210px;
+}
+
+.existingAccountWarning {
+  color: var(--error-text-color);
+  font-size: 11px;
+  line-height: 15px;
+  margin-top: 5px;
+}
+
+.fileInput {
+  padding-right: 10px;
+  margin-top: 8px;
+}
