--- app/components/views/LNPage/ConnectPage/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ConnectPage";
